Reduce the probability of false dead-locks on index entries

+++ enterprise/ha/src/main/java/org/neo4j/kernel/ha/lock/SlaveLocksClient.java
@@ -280,7 +280,7 @@ class SlaveLocksClient implements Locks.Client
     {
         makeSureTxHasBeenInitialized();
         return receiveLockResponse(
-                master.acquireExclusiveLock( requestContextFactory.newRequestContext( (int) getLockSessionId() ), resourceType, resourceId ));
+                master.acquireExclusiveLock( requestContextFactory.newRequestContext( getLockSessionId() ), resourceType, resourceId ));
     }
 
     private boolean receiveLockResponse( Response<LockResult> response )
